requirements.txt
google-adk==1.2.1
python-dotenv
pydantic
litellm==1.72.2
fastapi
uvicorn[standard]
asyncpg
langchain-mcp-adapters==0.1.1
langchain-community==0.3.24